rename: 각 도메인별 dto 디렉토리 생성
