+++ src/main/java/com/kl/grooveo/boundedContext/comment/dto/CommentForm.java
@@ -1,4 +1,4 @@
-package com.kl.grooveo.boundedContext.form;
+package com.kl.grooveo.boundedContext.comment.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
