+++ src/main/java/com/kl/grooveo/boundedContext/library/controller/FileUploadController.java
@@ -21,7 +21,7 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.kl.grooveo.base.rq.Rq;
-import com.kl.grooveo.boundedContext.form.SoundTrackForm;
+import com.kl.grooveo.boundedContext.library.dto.SoundTrackForm;
 import com.kl.grooveo.boundedContext.library.entity.FileInfo;
 import com.kl.grooveo.boundedContext.library.service.FileInfoService;
