+++ src/main/java/com/kl/grooveo/boundedContext/library/dto/FileInfoForm.java
@@ -1,4 +1,4 @@
-package com.kl.grooveo.boundedContext.form;
+package com.kl.grooveo.boundedContext.library.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
