+++ src/main/java/com/kl/grooveo/boundedContext/thumbsUp/controller/SoundThumbsUpController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.kl.grooveo.base.rq.Rq;
-import com.kl.grooveo.boundedContext.response.ThumbsUpDTO;
+import com.kl.grooveo.boundedContext.thumbsUp.dto.ThumbsUpDTO;
 import com.kl.grooveo.boundedContext.thumbsUp.service.SoundThumbsUpService;
 import com.kl.grooveo.boundedContext.thumbsUp.service.SoundThumbsUp_summaryService;
