+++ src/main/resources/application.yml
@@ -17,7 +17,7 @@ spring:
       enabled: true
   jpa:
     hibernate:
-      ddl-auto:
+      ddl-auto: create
     properties:
       hibernate:
         show_sql: true
